Add the Arquillian Suite Extension to improve performance
The Arquillian Suite Extension only allow a sinle deployment for the whole
test suite. The class to use for @Deployment scanning is defined in
arquillian.xml. The deployments will be deployed and undeployed in the
BeforeSuite and AfterSuite lifecycle events.

This extension is a temp work around until Arquillian Core supports sub
suites within a test suite.

Note that other Arquillian extensions that rely on manipulating the deployment
based on reading metadata from the TestClass will most likely not work
in combination with this suite extension.