Better way to detect thrown GAVAlreadyExistException (#426)
The exception is not directly thrown anymore. There are now few
guvnor-ala wrapper exceptions around.

This is not ideal at all. The try/catch block is being used to control
the program flow, instead of just checking exceptional states. However,
there does not seem to be an easy way to fix this without rewriting
large parts of the REST + build related code.

This fixes two failing tests in kie-wb-tests-rest:
 - ProjectIntegrationTest#testDeployProjectTwice
 - ProjectIntegrationTest#testInstallProjectTwice